+++ common/filepcx.h
@@ -29,7 +29,7 @@
 ;*                                                                         *
 ;*-------------------------------------------------------------------------*
 ;* Functions:                                                              *
-;* GraphicBufferClass* Read_PCX_File (char* name, BYTE* palette,void *buff, long size);
+;* GraphicBufferClass* Read_PCX_File (char* name, BYTE* palette,void *buff, int size);
 ;* GraphicBufferClass* Read_PCX_File (char* name, BYTE* palette, BufferClass& Buff);
 ;* int Write_PCX_File (char* name, GraphicViewPortClass& pic, BYTE* palette );*
 ;*= = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = =*/
@@ -70,7 +70,7 @@ typedef struct
 
 class GraphicBufferClass;
 
-GraphicBufferClass* Read_PCX_File(const char* name, char* palette = NULL, void* buff = NULL, long size = 0);
+GraphicBufferClass* Read_PCX_File(const char* name, char* palette = NULL, void* buff = NULL, int size = 0);
 int Write_PCX_File(char* name, GraphicViewPortClass& pic, unsigned char* palette);
 
 #endif
